+++ __tests__/main.spec.ts
@@ -10,7 +10,7 @@ import {alwaysFalseStateMock} from './classes/state-mock';
 test('processing an issue with no label will make it stale and close it, if it is old enough only if days-before-close is set to 0', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0
+    timeBeforeClose: 0
   };
   const TestIssueList: Issue[] = [
     generateIssue(opts, 1, 'An issue with no label', '2020-01-01T17:00:00Z')
@@ -35,7 +35,7 @@ test('processing an issue with no label and a start date as ECMAScript epoch in 
   const january2000 = 946681200000;
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2000.toString()
   };
   const TestIssueList: Issue[] = [
@@ -67,7 +67,7 @@ test('processing an issue with no label and a start date as ECMAScript epoch in 
   const january2021 = 1609455600000;
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2021.toString()
   };
   const TestIssueList: Issue[] = [
@@ -99,7 +99,7 @@ test('processing an issue with no label and a start date as ECMAScript epoch in 
   const january2000 = 946681200000000;
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2000.toString()
   };
   const TestIssueList: Issue[] = [
@@ -131,7 +131,7 @@ test('processing an issue with no label and a start date as ECMAScript epoch in 
   const january2021 = 1609455600000;
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2021.toString()
   };
   const TestIssueList: Issue[] = [
@@ -163,7 +163,7 @@ test('processing an issue with no label and a start date as ISO 8601 being befor
   const january2000 = '2000-01-01T00:00:00Z';
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2000.toString()
   };
   const TestIssueList: Issue[] = [
@@ -195,7 +195,7 @@ test('processing an issue with no label and a start date as ISO 8601 being after
   const january2021 = '2021-01-01T00:00:00Z';
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2021.toString()
   };
   const TestIssueList: Issue[] = [
@@ -227,7 +227,7 @@ test('processing an issue with no label and a start date as RFC 2822 being befor
   const january2000 = 'January 1, 2000 00:00:00';
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2000.toString()
   };
   const TestIssueList: Issue[] = [
@@ -259,7 +259,7 @@ test('processing an issue with no label and a start date as RFC 2822 being after
   const january2021 = 'January 1, 2021 00:00:00';
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 0,
+    timeBeforeClose: 0,
     startDate: january2021.toString()
   };
   const TestIssueList: Issue[] = [
@@ -289,8 +289,8 @@ test('processing an issue with no label and a start date as RFC 2822 being after
 test('processing an issue with no label will make it stale and close it, if it is old enough only if days-before-close is set to > 0 and days-before-issue-close is set to 0', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 1,
-    daysBeforeIssueClose: 0
+    timeBeforeClose: 1,
+    timeBeforeIssueClose: 0
   };
   const TestIssueList: Issue[] = [
     generateIssue(opts, 1, 'An issue with no label', '2020-01-01T17:00:00Z')
@@ -314,8 +314,8 @@ test('processing an issue with no label will make it stale and close it, if it i
 test('processing an issue with no label will make it stale and not close it, if it is old enough only if days-before-close is set to > 0 and days-before-issue-close is set to > 0', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 1,
-    daysBeforeIssueClose: 1
+    timeBeforeClose: 1,
+    timeBeforeIssueClose: 1
   };
   const TestIssueList: Issue[] = [
     generateIssue(opts, 1, 'An issue with no label', '2020-01-01T17:00:00Z')
@@ -338,7 +338,7 @@ test('processing an issue with no label will make it stale and not close it, if 
 test('processing an issue with no label will make it stale and not close it if days-before-close is set to > 0', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 15
+    timeBeforeClose: 15
   };
   const TestIssueList: Issue[] = [
     generateIssue(opts, 1, 'An issue with no label', '2020-01-01T17:00:00Z')
@@ -361,8 +361,8 @@ test('processing an issue with no label will make it stale and not close it if d
 test('processing an issue with no label will make it stale and not close it if days-before-close is set to -1 and days-before-issue-close is set to > 0', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: -1,
-    daysBeforeIssueClose: 15
+    timeBeforeClose: -1,
+    timeBeforeIssueClose: 15
   };
   const TestIssueList: Issue[] = [
     generateIssue(opts, 1, 'An issue with no label', '2020-01-01T17:00:00Z')
@@ -386,7 +386,7 @@ test('processing an issue with no label will not make it stale if days-before-st
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
     staleIssueMessage: '',
-    daysBeforeStale: -1
+    timeBeforeStale: -1
   };
   const TestIssueList: Issue[] = [
     generateIssue(opts, 1, 'An issue with no label', '2020-01-01T17:00:00Z')
@@ -410,8 +410,8 @@ test('processing an issue with no label will not make it stale if days-before-st
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
     staleIssueMessage: '',
-    daysBeforeStale: -1,
-    daysBeforeIssueStale: -1
+    timeBeforeStale: -1,
+    timeBeforeIssueStale: -1
   };
   const TestIssueList: Issue[] = [
     generateIssue(opts, 1, 'An issue with no label', '2020-01-01T17:00:00Z')
@@ -462,7 +462,7 @@ test('processing an issue with no label will make it stale but not close it', as
 test('processing a stale issue will close it', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 30
+    timeBeforeClose: 30
   };
   const TestIssueList: Issue[] = [
     generateIssue(
@@ -558,8 +558,8 @@ test('processing a stale issue containing a slash in the label will close it', a
 test('processing a stale issue will close it when days-before-issue-stale override days-before-stale', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 30,
-    daysBeforeIssueStale: 30
+    timeBeforeClose: 30,
+    timeBeforeIssueStale: 30
   };
   const TestIssueList: Issue[] = [
     generateIssue(
@@ -591,7 +591,7 @@ test('processing a stale issue will close it when days-before-issue-stale overri
 test('processing a stale PR will close it', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 30
+    timeBeforeClose: 30
   };
   const TestIssueList: Issue[] = [
     generateIssue(
@@ -623,8 +623,8 @@ test('processing a stale PR will close it', async () => {
 test('processing a stale PR will close it when days-before-pr-stale override days-before-stale', async () => {
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeClose: 30,
-    daysBeforePrClose: 30
+    timeBeforeClose: 30,
+    timeBeforePrClose: 30
   };
   const TestIssueList: Issue[] = [
     generateIssue(
@@ -656,7 +656,7 @@ test('processing a stale PR will close it when days-before-pr-stale override day
 test('processing a stale issue will close it even if configured not to mark as stale', async () => {
   const opts = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: -1,
+    timeBeforeStale: -1,
     staleIssueMessage: ''
   };
   const TestIssueList: Issue[] = [
@@ -689,8 +689,8 @@ test('processing a stale issue will close it even if configured not to mark as s
 test('processing a stale issue will close it even if configured not to mark as stale when days-before-issue-stale override days-before-stale', async () => {
   const opts = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 0,
-    daysBeforeIssueStale: -1,
+    timeBeforeStale: 0,
+    timeBeforeIssueStale: -1,
     staleIssueMessage: ''
   };
   const TestIssueList: Issue[] = [
@@ -723,7 +723,7 @@ test('processing a stale issue will close it even if configured not to mark as s
 test('processing a stale PR will close it even if configured not to mark as stale', async () => {
   const opts = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: -1,
+    timeBeforeStale: -1,
     stalePrMessage: ''
   };
   const TestIssueList: Issue[] = [
@@ -756,8 +756,8 @@ test('processing a stale PR will close it even if configured not to mark as stal
 test('processing a stale PR will close it even if configured not to mark as stale when days-before-pr-stale override days-before-stale', async () => {
   const opts = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 0,
-    daysBeforePrStale: -1,
+    timeBeforeStale: 0,
+    timeBeforePrStale: -1,
     stalePrMessage: ''
   };
   const TestIssueList: Issue[] = [
@@ -1161,7 +1161,7 @@ test('exempt pr labels will not be marked stale', async () => {
 
 test('stale issues should not be closed if days is set to -1', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeClose = -1;
+  opts.timeBeforeClose = -1;
   const TestIssueList: Issue[] = [
     generateIssue(
       opts,
@@ -1412,8 +1412,8 @@ test('stale label containing a space should be removed if a comment was added to
 
 test('stale issues should not be closed until after the closed number of days', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeStale = 5; // stale after 5 days
-  opts.daysBeforeClose = 1; // closes after 6 days
+  opts.timeBeforeStale = 5; // stale after 5 days
+  opts.timeBeforeClose = 1; // closes after 6 days
   const lastUpdate = new Date();
   lastUpdate.setDate(lastUpdate.getDate() - 5);
   const TestIssueList: Issue[] = [
@@ -1444,8 +1444,8 @@ test('stale issues should not be closed until after the closed number of days', 
 
 test('stale issues should be closed if the closed nubmer of days (additive) is also passed', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeStale = 5; // stale after 5 days
-  opts.daysBeforeClose = 1; // closes after 6 days
+  opts.timeBeforeStale = 5; // stale after 5 days
+  opts.timeBeforeClose = 1; // closes after 6 days
   const lastUpdate = new Date();
   lastUpdate.setDate(lastUpdate.getDate() - 7);
   const TestIssueList: Issue[] = [
@@ -1478,8 +1478,8 @@ test('stale issues should be closed if the closed nubmer of days (additive) is a
 
 test('stale issues should not be closed until after the closed number of days (long)', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeStale = 5; // stale after 5 days
-  opts.daysBeforeClose = 20; // closes after 25 days
+  opts.timeBeforeStale = 5; // stale after 5 days
+  opts.timeBeforeClose = 20; // closes after 25 days
   const lastUpdate = new Date();
   lastUpdate.setDate(lastUpdate.getDate() - 10);
   const TestIssueList: Issue[] = [
@@ -1510,8 +1510,8 @@ test('stale issues should not be closed until after the closed number of days (l
 
 test('skips stale message on issues when stale-issue-message is empty', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeStale = 5; // stale after 5 days
-  opts.daysBeforeClose = 20; // closes after 25 days
+  opts.timeBeforeStale = 5; // stale after 5 days
+  opts.timeBeforeClose = 20; // closes after 25 days
   opts.staleIssueMessage = '';
   const lastUpdate = new Date();
   lastUpdate.setDate(lastUpdate.getDate() - 10);
@@ -1555,8 +1555,8 @@ test('skips stale message on issues when stale-issue-message is empty', async ()
 
 test('send stale message on issues when stale-issue-message is not empty', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeStale = 5; // stale after 5 days
-  opts.daysBeforeClose = 20; // closes after 25 days
+  opts.timeBeforeStale = 5; // stale after 5 days
+  opts.timeBeforeClose = 20; // closes after 25 days
   opts.staleIssueMessage = 'dummy issue message';
   const lastUpdate = new Date();
   lastUpdate.setDate(lastUpdate.getDate() - 10);
@@ -1600,8 +1600,8 @@ test('send stale message on issues when stale-issue-message is not empty', async
 
 test('skips stale message on prs when stale-pr-message is empty', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeStale = 5; // stale after 5 days
-  opts.daysBeforeClose = 20; // closes after 25 days
+  opts.timeBeforeStale = 5; // stale after 5 days
+  opts.timeBeforeClose = 20; // closes after 25 days
   opts.stalePrMessage = '';
   const lastUpdate = new Date();
   lastUpdate.setDate(lastUpdate.getDate() - 10);
@@ -1646,8 +1646,8 @@ test('skips stale message on prs when stale-pr-message is empty', async () => {
 
 test('send stale message on prs when stale-pr-message is not empty', async () => {
   const opts = {...DefaultProcessorOptions};
-  opts.daysBeforeStale = 5; // stale after 5 days
-  opts.daysBeforeClose = 20; // closes after 25 days
+  opts.timeBeforeStale = 5; // stale after 5 days
+  opts.timeBeforeClose = 20; // closes after 25 days
   opts.stalePrMessage = 'dummy pr message';
   const lastUpdate = new Date();
   lastUpdate.setDate(lastUpdate.getDate() - 10);
@@ -2037,8 +2037,8 @@ test('processing an issue opened since 2 days and with the option "daysBeforeIss
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 3
+    timeBeforeStale: 10,
+    timeBeforeIssueStale: 3
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);
@@ -2064,8 +2064,8 @@ test('processing an issue opened since 2 days and with the option "daysBeforeIss
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 2
+    timeBeforeStale: 10,
+    timeBeforeIssueStale: 2
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);
@@ -2091,8 +2091,8 @@ test('processing an issue opened since 2 days and with the option "daysBeforeIss
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 1
+    timeBeforeStale: 10,
+    timeBeforeIssueStale: 1
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);
@@ -2118,8 +2118,8 @@ test('processing an issue opened since 1 hour and with the option "daysBeforeIss
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 0.1666666667
+    timeBeforeStale: 10,
+    timeBeforeIssueStale: 0.1666666667
   };
   const issueDate = new Date();
   issueDate.setHours(issueDate.getHours() - 1);
@@ -2145,8 +2145,8 @@ test('processing an issue opened since 4 hours and with the option "daysBeforeIs
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 0.1666666667
+    timeBeforeStale: 10,
+    timeBeforeIssueStale: 0.1666666667
   };
   const issueDate = new Date();
   issueDate.setHours(issueDate.getHours() - 4);
@@ -2172,8 +2172,8 @@ test('processing an issue opened since 5 hours and with the option "daysBeforeIs
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 0.1666666667
+    timeBeforeStale: 10,
+    timeBeforeIssueStale: 0.1666666667
   };
   const issueDate = new Date();
   issueDate.setHours(issueDate.getHours() - 5);
@@ -2199,8 +2199,8 @@ test('processing a pull request opened since 2 days and with the option "daysBef
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 3
+    timeBeforeStale: 10,
+    timeBeforePrStale: 3
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);
@@ -2234,8 +2234,8 @@ test('processing a pull request opened since 2 days and with the option "daysBef
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 2
+    timeBeforeStale: 10,
+    timeBeforePrStale: 2
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);
@@ -2269,8 +2269,8 @@ test('processing a pull request opened since 2 days and with the option "daysBef
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 1
+    timeBeforeStale: 10,
+    timeBeforePrStale: 1
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);
@@ -2304,8 +2304,8 @@ test('processing a pull request opened since 1 hour and with the option "daysBef
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 0.1666666667
+    timeBeforeStale: 10,
+    timeBeforePrStale: 0.1666666667
   };
   const issueDate = new Date();
   issueDate.setHours(issueDate.getHours() - 1);
@@ -2339,8 +2339,8 @@ test('processing a pull request opened since 4 hours and with the option "daysBe
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 0.1666666667
+    timeBeforeStale: 10,
+    timeBeforePrStale: 0.1666666667
   };
   const issueDate = new Date();
   issueDate.setHours(issueDate.getHours() - 4);
@@ -2374,8 +2374,8 @@ test('processing a pull request opened since 5 hours and with the option "daysBe
   expect.assertions(2);
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 0.1666666667
+    timeBeforeStale: 10,
+    timeBeforePrStale: 0.1666666667
   };
   const issueDate = new Date();
   issueDate.setHours(issueDate.getHours() - 5);
@@ -2521,8 +2521,8 @@ test('processing an issue stale since less than the daysBeforeStale with a stale
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
     staleIssueLabel: 'stale-label',
-    daysBeforeStale: 30,
-    daysBeforeClose: 7,
+    timeBeforeStale: 30,
+    timeBeforeClose: 7,
     closeIssueMessage: 'close message',
     removeStaleWhenUpdated: false
   };
@@ -2564,8 +2564,8 @@ test('processing an issue stale since less than the daysBeforeStale without a st
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
     staleIssueLabel: 'stale-label',
-    daysBeforeStale: 30,
-    daysBeforeClose: 7,
+    timeBeforeStale: 30,
+    timeBeforeClose: 7,
     closeIssueMessage: 'close message',
     removeStaleWhenUpdated: false
   };
@@ -2606,8 +2606,8 @@ test('processing a pull request to be stale with the "stalePrMessage" option set
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
     stalePrMessage: 'This PR is stale',
-    daysBeforeStale: 10,
-    daysBeforePrStale: 1
+    timeBeforeStale: 10,
+    timeBeforePrStale: 1
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);
@@ -2643,8 +2643,8 @@ test('processing a pull request to be stale with the "stalePrMessage" option set
   const opts: IIssuesProcessorOptions = {
     ...DefaultProcessorOptions,
     stalePrMessage: '',
-    daysBeforeStale: 10,
-    daysBeforePrStale: 1
+    timeBeforeStale: 10,
+    timeBeforePrStale: 1
   };
   const issueDate = new Date();
   issueDate.setDate(issueDate.getDate() - 2);